requirements.txt
langchain 
langchain-groq
langchain-community
langchain-huggingface
requests
streamlit 
python-dotenv
PyPDF2
python-docx
chromadb==0.4.24
sentence-transformers
faiss-gpu
faiss-cpu
unstructured
unstructured[pdf]
pydantic
langchain-unstructured

